Implement basic server and test probe (#12)
Implement the server as a "composer" crate for a (IMO) better fitting
name. In addition to the binary give it a library section that exposes
logic reused by probes as well as the possible event types.

Start with very basic sound effect of just playing a sample from a file.
This needs some improvements. But we gotta start somewhere.

Also implement a basic test probe that sends test events in a pulsating
frequency.