Use the same family for local as for remote address
To construct a socket it must be first bound to a local address and only
then it can connect to a remote address.

For the connection to succeed both addresses must be of the same family
(IPv4 or 6).

When constructing the client first parse the given server address to
determine its family and then bind to a wildcard address of the same
family.

fixes #27