Add burst mode to test probe
Add a mode to the test probe for testing event ordering and
time-positioning of the composer in a controlled way.

Make sure the sequence of event timestamps progresses in a stable way
independent of the bursting rhythm which tends to hiccup. (Probably due
to varying execution times and imprecise thread::sleep)